# docs/architecture/api_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Daily Stock Analysis API",
    "description": "A股/港股/美股自选股智能分析系统 API\n\n## 功能模块\n- 股票分析：触发 AI 智能分析\n- 历史记录：查询历史分析报告\n- 股票数据：获取行情数据\n\n## 认证方式\n支持可选管理员认证：ADMIN_AUTH_ENABLED=true 时，除登录、状态、健康检查和 OpenAPI 文档外，/api/v1/* 需要有效管理员会话 Cookie；关闭时不强制认证。",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "本地开发服务器"
    }
  ],
...
  "paths": {
...
    "/api/v1/calendar/events": {
      "post": {
        "tags": [
          "Calendar"
        ],
        "summary": "Create a user calendar event",
        "operationId": "create_event_api_v1_calendar_events_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarEventCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarEventItem"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Calendar"
        ],
        "summary": "List upcoming calendar events",
        "description": "Shared inclusive date-range query for Dashboard and symbol consumers. Defaults to seven calendar dates starting today.",
        "operationId": "list_events_api_v1_calendar_events_get",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          },
          {
            "name": "scope_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "market",
                    "symbol",
                    "portfolio",
                    "sector",
                    "custom"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scope Type"
            }
          },
          {
            "name": "scope_value",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scope Value"
            }
          },
          {
            "name": "market",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "cn",
                    "hk",
                    "us",
                    "jp",
                    "kr",
                    "tw",
                    "global"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Market"
            }
          },
          {
            "name": "symbol",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 32
                },
                {
                  "type": "null"
                }
              ],
              "title": "Symbol"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 100,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarEventListResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/calendar/events/{event_id}": {
      "delete": {
        "tags": [
          "Calendar"
        ],
        "summary": "Delete a user calendar event",
        "operationId": "delete_event_api_v1_calendar_events__event_id__delete",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarEventDeleteResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "错误类型",
            "example": "validation_error"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "错误详情",
            "example": "请求参数错误"
          },
          "detail": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Detail",
            "description": "附加错误信息"
          }
        },
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorResponse",
        "description": "错误响应",
        "example": {
          "error": "not_found",
          "message": "资源不存在"
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "CalendarCoverageSummary": {
        "properties": {
          "status": {
            "type": "string",
            "const": "manual_only",
            "title": "Status"
          },
          "external_sources_configured": {
            "type": "boolean",
            "title": "External Sources Configured"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "status",
          "external_sources_configured",
          "message"
        ],
        "title": "CalendarCoverageSummary"
      },
      "CalendarEventCreateRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "event_type": {
            "type": "string",
            "enum": [
              "earnings",
              "dividend",
              "lockup_unlock",
              "macro",
              "user",
              "monitor"
            ],
            "title": "Event Type"
          },
          "scope_type": {
            "type": "string",
            "enum": [
              "market",
              "symbol",
              "portfolio",
              "sector",
              "custom"
            ],
            "title": "Scope Type"
          },
          "scope_value": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope Value"
          },
          "market": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "cn",
                  "hk",
                  "us",
                  "jp",
                  "kr",
                  "tw",
                  "global"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Market"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          },
          "event_date": {
            "type": "string",
            "format": "date",
            "title": "Event Date"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "title",
          "event_type",
          "scope_type",
          "event_date"
        ],
        "title": "CalendarEventCreateRequest"
      },
      "CalendarEventDeleteResponse": {
        "properties": {
          "deleted": {
            "type": "integer",
            "title": "Deleted"
          }
        },
        "type": "object",
        "required": [
          "deleted"
        ],
        "title": "CalendarEventDeleteResponse"
      },
      "CalendarEventItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type"
          },
          "scope_type": {
            "type": "string",
            "title": "Scope Type"
          },
          "scope_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope Value"
          },
          "market": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          },
          "event_date": {
            "type": "string",
            "title": "Event Date"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "coverage_status": {
            "type": "string",
            "title": "Coverage Status"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "event_type",
          "scope_type",
          "event_date",
          "source",
          "coverage_status"
        ],
        "title": "CalendarEventItem"
      },
      "CalendarEventListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CalendarEventItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "start_date": {
            "type": "string",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "title": "End Date"
          },
          "coverage": {
            "$ref": "#/components/schemas/CalendarCoverageSummary"
          }
        },
        "type": "object",
        "required": [
          "total",
          "page",
          "page_size",
          "start_date",
          "end_date",
          "coverage"
        ],
        "title": "CalendarEventListResponse"
      }
    },
    "securitySchemes": {
      "AdminSessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "dsa_session"
      }
    }
  }
}